Add hover and selected effects for server/channel/members

+++ frontend/reducers/selectors.js
@@ -25,7 +25,7 @@ export const selectChannels = (state, serverId) => {
 export const selectChannel = (state, serverId, ccId) => {
   const channelIds = state.entities.servers[serverId]?.channels;
 
-  if (channelIds.includes(parseFloat(ccId))) {
+  if (channelIds.includes(parseInt(ccId))) {
     return state.entities.channels[ccId];
   } else {
     return state.entities.channels[channelIds[0]];
